Move data source logic to repository
