+++ src/domain/auth/auth.repository.interface.ts
@@ -6,10 +6,12 @@ export const IAuthRepository = Symbol('IAuthRepository');
 export interface IAuthRepository {
   generateNonce(): Promise<{ nonce: string }>;
 
-  isAuthorized(args: {
-    message: SiweMessage;
-    signature: `0x${string}`;
-  }): Promise<boolean>;
+  verify(args: { message: SiweMessage; signature: `0x${string}` }): Promise<{
+    accessToken: string;
+    tokenType: string;
+    notBefore: number | null;
+    expiresIn: number | null;
+  }>;
 
   getAccessToken(request: Request, tokenType: string): string | null;
 }
