+++ src/domain/auth/auth.repository.ts
@@ -2,13 +2,16 @@ import { CacheRouter } from '@/datasources/cache/cache.router';
 import { IAuthRepository } from '@/domain/auth/auth.repository.interface';
 import { SiweMessage } from '@/domain/auth/entities/siwe-message.entity';
 import { IAuthApi } from '@/domain/interfaces/auth-api.interface';
-import { Inject, Injectable } from '@nestjs/common';
+import { Inject, Injectable, UnauthorizedException } from '@nestjs/common';
 import { Request } from 'express';
 import { IConfigurationService } from '@/config/configuration.service.interface';
 import {
   CacheService,
   ICacheService,
 } from '@/datasources/cache/cache.service.interface';
+import { AuthService } from '@/routes/auth/auth.service';
+import { VerifyAuthMessageDto } from '@/routes/auth/entities/schemas/verify-auth-message.dto';
+import { IJwtService } from '@/datasources/jwt/jwt.service.interface';
 
 @Injectable()
 export class AuthRepository implements IAuthRepository {
@@ -20,6 +23,8 @@ export class AuthRepository implements IAuthRepository {
     @Inject(IConfigurationService)
     private readonly configurationService: IConfigurationService,
     @Inject(CacheService) private readonly cacheService: ICacheService,
+    @Inject(IJwtService)
+    private readonly jwtService: IJwtService,
   ) {
     this.nonceTtlInSeconds = this.configurationService.getOrThrow(
       'auth.nonceTtlSeconds',
@@ -44,6 +49,67 @@ export class AuthRepository implements IAuthRepository {
   }
 
   /**
+   * Verifies the validity of a signed message and returns an access token:
+   *
+   * 1. Ensure the message itself has not expired.
+   * 2. Ensure the nonce was generated by us/is not a replay attack.
+   * 3. Verify the signature of the message.
+   * 4. Return an access token if all checks pass.
+   *
+   * @param args - DTO containing the message and signature to verify.
+   *
+   * The following adhere to JWT standard {@link https://datatracker.ietf.org/doc/html/rfc7519}
+   *
+   * @returns accessToken - JWT access token
+   * @returns tokenType - token type ('Bearer') to be used in the `Authorization` header
+   * @returns notBefore - epoch from when token is valid (if applicable, otherwise null)
+   * @returns expiresIn - time in seconds until the token expires (if applicable, otherwise null)
+   */
+  async verify(args: VerifyAuthMessageDto): Promise<{
+    accessToken: string;
+    tokenType: string;
+    notBefore: number | null;
+    expiresIn: number | null;
+  }> {
+    const isAuthorized = await this.isAuthorized(args).catch(() => false);
+
+    if (!isAuthorized) {
+      throw new UnauthorizedException();
+    }
+
+    const dateWhenTokenIsValid = args.message.notBefore
+      ? new Date(args.message.notBefore)
+      : null;
+    const dateWhenTokenExpires = args.message.expirationTime
+      ? new Date(args.message.expirationTime)
+      : null;
+
+    const secondsUntilTokenIsValid = dateWhenTokenIsValid
+      ? this.getSecondsUntil(dateWhenTokenIsValid)
+      : null;
+    const secondsUntilTokenExpires = dateWhenTokenExpires
+      ? this.getSecondsUntil(dateWhenTokenExpires)
+      : null;
+
+    const accessToken = this.jwtService.sign(args.message, {
+      ...(secondsUntilTokenIsValid !== null && {
+        notBefore: secondsUntilTokenIsValid,
+      }),
+      ...(secondsUntilTokenExpires !== null && {
+        expiresIn: secondsUntilTokenExpires,
+      }),
+    });
+
+    return {
+      tokenType: AuthService.AUTH_TOKEN_TOKEN_TYPE,
+      accessToken,
+      // Differing measurements match JWT standard {@link https://datatracker.ietf.org/doc/html/rfc7519}
+      notBefore: dateWhenTokenIsValid?.getTime() ?? null,
+      expiresIn: secondsUntilTokenExpires,
+    };
+  }
+
+  /**
    * Verifies that a message is valid according to its expiration date,
    * signature and nonce.
    *
@@ -52,7 +118,7 @@ export class AuthRepository implements IAuthRepository {
    *
    * @returns boolean - whether the message is valid
    */
-  async isAuthorized(args: {
+  private async isAuthorized(args: {
     message: SiweMessage;
     signature: `0x${string}`;
   }): Promise<boolean> {
@@ -73,6 +139,10 @@ export class AuthRepository implements IAuthRepository {
     await this.cacheService.deleteByKey(cacheDir.key);
 
     return !isExpired && isValidSignature && isValidNonce;
+  }
+
+  private getSecondsUntil(date: Date): number {
+    return Math.floor((date.getTime() - Date.now()) / 1_000);
   }
 
   /**
